Add an option to collect profiling data using cProfile module
In order to collect profiling data, add the '--profiling' argument to the
yolo_video.py command line (or to its 'run' configuration).
If '--profiling_stats_file=<some_file>' argument is not provided, the profiling
statistics will be printed to screen at the end of execution.
If '--profiling_stats_file=<some_file>' argument is provided, the '--profiling'
flag become redundant.